Add support for music collections
Include audio playlists when searching, and allow tracks to be added to
collections. Additionally allow listed playlists to be filtered by type
to help reduce the number of playlists we display when the user doesn't
enter a specific value.

Also mentioned new Python version requirement, 3.10, for use of
match/case, and bump requirements.txt.